refactor:adjust the api_server directory name, change api_server to apiserver, issue #1727
