+++ src/apiserver/logics/v2/common/converter/converter.go
@@ -21,8 +21,8 @@ import (
 	"strings"
 	"time"
 
-	"configcenter/src/api_server/logics/v2/common/defs"
-	"configcenter/src/api_server/logics/v2/common/utils"
+	"configcenter/src/apiserver/logics/v2/common/defs"
+	"configcenter/src/apiserver/logics/v2/common/utils"
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	ccError "configcenter/src/common/errors"
