+++ src/apiserver/service/v2/application.go
@@ -18,9 +18,9 @@ import (
 	"strconv"
 	"strings"
 
-	"configcenter/src/api_server/logics/v2/common/converter"
-	"configcenter/src/api_server/logics/v2/common/defs"
-	"configcenter/src/api_server/logics/v2/common/utils"
+	"configcenter/src/apiserver/logics/v2/common/converter"
+	"configcenter/src/apiserver/logics/v2/common/defs"
+	"configcenter/src/apiserver/logics/v2/common/utils"
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"configcenter/src/common/mapstr"
