+++ src/apiserver/service/v2/gse.go
@@ -17,8 +17,8 @@ import (
 
 	"github.com/emicklei/go-restful"
 
-	"configcenter/src/api_server/logics/v2/common/converter"
-	"configcenter/src/api_server/logics/v2/common/utils"
+	"configcenter/src/apiserver/logics/v2/common/converter"
+	"configcenter/src/apiserver/logics/v2/common/utils"
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"configcenter/src/common/util"
