+++ src/apiserver/service/v2/search.go
@@ -19,9 +19,9 @@ import (
 	"strconv"
 	"strings"
 
-	"configcenter/src/api_server/logics/v2/common/converter"
-	"configcenter/src/api_server/logics/v2/common/defs"
-	"configcenter/src/api_server/logics/v2/common/utils"
+	"configcenter/src/apiserver/logics/v2/common/converter"
+	"configcenter/src/apiserver/logics/v2/common/defs"
+	"configcenter/src/apiserver/logics/v2/common/utils"
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	ccErr "configcenter/src/common/errors"
