+++ src/apiserver/service/v2/userapi.go
@@ -19,7 +19,7 @@ import (
 	"strconv"
 	"strings"
 
-	"configcenter/src/api_server/logics/v2/common/converter"
+	"configcenter/src/apiserver/logics/v2/common/converter"
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"configcenter/src/common/mapstr"
